Setting up LogRotateFs.
This commit introduces the LogRotateFs class and its private
implementation.

The most part of the old interface (the one of LogRotate) has been kept
but using (for the most part) the std::filesystem::path.

The doxygen documentation has been completely remade. This commit only
introduces the interface of LogRotateFs without its implementation.

The implementation will come later (with unit testing).